pyproject.toml
[project]
name = "frigate-tools"
version = "0.1.0"
description = "CLI utilities for working with Frigate NVR recordings"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opentelemetry-api>=1.28.0",
    "opentelemetry-exporter-otlp>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "rich>=14.2.0",
    "structlog>=24.4.0",
    "typer>=0.20.0",
]

[project.scripts]
frigate-tools = "frigate_tools.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/frigate_tools"]
